requirements.txt
langchain
sentence-transformers
chromadb
llama-cpp-python
huggingface_hub
pymupdf
python-dotenv
langchain-community

